fixed table creation script(price data type to decimal(10,2)

+++ src/main/resources/init_db.sql
@@ -2,5 +2,5 @@ CREATE TABLE books
 (
     id    BIGINT AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(255) NOT NULL,
-    price BIGINT
+    price DECIMAL(10, 2)
 );
